refactor: move timeslot logic into timepicker
